Fold the CallNodeInfoInverted interface into the implementation class.
