+++ src/components/timeline/TrackThread.js
@@ -52,13 +52,13 @@ import type {
   IndexIntoSamplesTable,
   Milliseconds,
   StartEndRange,
-  CallNodeInfo,
   ImplementationFilter,
   IndexIntoCallNodeTable,
   SelectedState,
   State,
   ThreadsKey,
 } from 'firefox-profiler/types';
+import type { CallNodeInfo } from 'firefox-profiler/profile-logic/call-node-info';
 
 import type { ConnectedProps } from 'firefox-profiler/utils/connect';
