+++ src/profile-logic/line-timings.js
@@ -9,8 +9,6 @@ import type {
   FuncTable,
   StackTable,
   SamplesLikeTable,
-  CallNodeInfo,
-  CallNodeInfoInverted,
   IndexIntoCallNodeTable,
   IndexIntoStringTable,
   StackLineInfo,
@@ -19,6 +17,7 @@ import type {
 } from 'firefox-profiler/types';
 
 import { getMatchingAncestorStackForInvertedCallNode } from './profile-data';
+import type { CallNodeInfo, CallNodeInfoInverted } from './call-node-info';
 
 /**
  * For each stack in `stackTable`, and one specific source file, compute the
